% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a_priori_delimiter_ranks}
\alias{a_priori_delimiter_ranks}
\title{A table with common file delimiters and their ranking according to popularity}
\format{A tibble with columns:
\describe{
\item{char_raw}{Binary representation of delimiter}
\item{rank}{Rank of corresponding delimiter}
}}
\usage{
a_priori_delimiter_ranks
}
\description{
A table with common file delimiters and their ranking according to popularity
}
\keyword{datasets}
