% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_aux.R
\name{count_chars}
\alias{count_chars}
\title{Auxiliar function to count characters in a line}
\usage{
count_chars(line)
}
\arguments{
\item{line}{A string of characters}
}
\description{
Auxiliar function to count characters in a line
}
