% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enum_repeated_names}
\alias{enum_repeated_names}
\title{Enumarate repeated names}
\usage{
enum_repeated_names(str)
}
\arguments{
\item{str}{a character vector.}
}
\description{
Enumarate repeated names and put it as suffix to make every name from a vector unique.
}
\examples{

enum_repeated_names(c("a", "a", "b", "c", "c", "c", "d"))

}
\seealso{
\code{\link[=tide_names]{tide_names()}}
}
