% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frk.R
\name{frk_read_delim}
\alias{frk_read_delim}
\title{Read flat file knowing nothing about them}
\usage{
frk_read_delim(file, delim = guess_delim(file, guess_max)$char[1],
  quote = guess_quote(file, guess_max), escape_backslash = FALSE,
  escape_double = TRUE, col_names = guess_has_header(file, guess_max),
  col_types = NULL, locale = readr::locale(encoding = guess_encoding(file,
  guess_max)$encoding[1], decimal_mark = guess_decimal_mark(file, guess_max),
  grouping_mark = guess_grouping_mark(file, guess_max)), na = c("", "NA"),
  quoted_na = TRUE, comment = "", trim_ws = TRUE,
  skip = guess_skip(file, guess_max), n_max = Inf, guess_max = min(10,
  n_max))
}
\arguments{
\item{file}{Path to file}

\item{delim}{(Guessed) Delimiter of file}

\item{quote}{(Guessed) Character used to quote strings}

\item{escape_backslash}{Whether backslashed are used to escape special characters}

\item{escape_double}{Whether quotes are escaped by doubling them}

\item{col_names}{(Guessed) Whether file has column names or vector with column names}

\item{col_types}{Vector with types to be forced on columns or NULL for guess}

\item{locale}{(Guessed) Some region-dependent information}

\item{na}{String or character vector to use for missing values}

\item{quoted_na}{Whether missing values can be quoted}

\item{comment}{String that indicates comments}

\item{trim_ws}{Whether to trim whitespace from each field}

\item{skip}{(Guessed) Number of lines to skip before reading data}

\item{n_max}{Maximum number of records to read}

\item{guess_max}{Maximum number of records to use for guess}
}
\description{
This is a wrapper for \code{\link[readr:read_delim]{readr::read_delim()}} that guesses the main arguments,
namely \code{delim}, \code{quote}, \code{col_names}, \code{encoding}, \code{decimal_mark}, \code{grouping_mark}, and \code{skip}.
}
\examples{
\dontrun{
# Create sample file
file <- tempfile()
write.table(
  dplyr::storms, file, sep = "|", dec = ",",
  col.names = TRUE, row.names = FALSE)

# Read file
tbl <- frk_read_delim(file)
glimpse(tbl)
}

}
\seealso{
\code{\link[readr:read_delim]{readr::read_delim()}}, \code{\link[readr:locale]{readr::locale()}}
}
