% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frk.R
\name{frk_stack}
\alias{frk_stack}
\title{Stack a set of files from a instruction}
\usage{
frk_stack(instruction)
}
\arguments{
\item{file}{A tibble returned by \code{frk_summarise()}.}
}
\description{
This is a wrapper for \code{\link[readr:read_delim]{readr::read_delim()}} that guesses the main arguments,
namely \code{delim}, \code{quote}, \code{col_names}, \code{encoding}, \code{decimal_mark}, \code{grouping_mark}, and \code{skip}.
}
\examples{
\dontrun{
dir <- "~/folder_full_of_flat_files"
stacked_files <- dir \%>\% frk_summarise \%>\% frk_stack
}

}
\seealso{
\code{\link[=frk_summarise]{frk_summarise()}}, \code{\link[=frk_read_delim]{frk_read_delim()}}
}
