% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frk.R
\name{frk_summarise}
\alias{frk_summarise}
\title{Guess all structural characteristics of one or more flat files}
\usage{
frk_summarise(files, guess_max = 10, verbose = FALSE)
}
\arguments{
\item{files}{Path to file or character vector of paths to files}

\item{guess_max}{Maximum number of records to use for guess}

\item{verbose}{Whether to output guess as message}
}
\value{
A list of lists with the guesses for each file
}
\description{
Runs all functions from the \code{guess} family to gather all structural
characteristics of given file(s).
}
\examples{
\dontrun{
# Create sample files
file <- tempfile()
write.table(
  dplyr::storms, file, sep = "|", dec = ",",
  col.names = TRUE, row.names = FALSE)
file2 <- tempfile()
write_csv(dplyr::storms, file2)

# Run all guesses
summ <- frk_summarise(file)
str(summ)

# Run all guesses on multiple files
summ <- frk_summarise(c(file, file2))
str(summ)
}

}
\seealso{
\code{\link[=guess_delim]{guess_delim()}}, \code{\link[=guess_col_names]{guess_col_names()}}, and the whole \code{guess} family
}
