% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frk.R
\name{frk_summarise_}
\alias{frk_summarise_}
\title{Guess all structural characteristics of a flat file}
\usage{
frk_summarise_(file, guess_max = 10, skip = NULL, verbose = FALSE)
}
\arguments{
\item{file}{Path to file}

\item{guess_max}{Maximum number of records to use for guess}

\item{verbose}{Whether to output guess as message}
}
\value{
A named list with the guesses for the specified file
}
\description{
Guess all structural characteristics of a flat file
}
