% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{guess}
\alias{guess}
\alias{guess_delim}
\alias{guess_encoding}
\alias{guess_has_header}
\alias{guess_col_types}
\alias{guess_col_names}
\alias{guess_quote}
\alias{guess_skip}
\alias{guess_decimal_mark}
\alias{guess_grouping_mark}
\title{Guess structural characteristic of a flat file}
\usage{
guess_delim(file, guess_max = 10, verbose = FALSE,
  encoding = guess_encoding(file, guess_max), skip = guess_skip(file,
  guess_max))

guess_encoding(file, guess_max = 10, verbose = FALSE, threshold = 0.2)

guess_has_header(file, guess_max = 10, verbose = FALSE)

guess_col_types(file, guess_max = 10, verbose = FALSE,
  delim = guess_delim(file, guess_max)$char[1], skip = guess_skip(file,
  guess_max), encoding = guess_encoding(file, guess_max))

guess_col_names(file, guess_max = 10, verbose = FALSE,
  delim = guess_delim(file, guess_max)$char[1],
  header = guess_has_header(file, guess_max), quote = guess_quote(file,
  guess_max), encoding = guess_encoding(file, guess_max),
  skip = guess_skip(file, guess_max))

guess_quote(file, guess_max = 10, verbose = FALSE, skip = guess_skip(file,
  guess_max))

guess_skip(file, guess_max = 10, verbose = FALSE)

guess_decimal_mark(file, guess_max = 10, verbose = FALSE,
  delim = guess_delim(file, guess_max)$char[1], quote = guess_quote(file,
  guess_max), skip = guess_skip(file, guess_max))

guess_grouping_mark(file, guess_max = 10, verbose = FALSE)
}
\arguments{
\item{file}{Path to file}

\item{guess_max}{Maximum number of records to use for guess}

\item{verbose}{Whether to output guess as message}

\item{encoding}{Default encoding. This only affects how the file is read. Guessed if not specified.}

\item{skip}{Number of lines to skip before reading data. Guessed if not specified.}
}
\value{
All \code{guess} functions return an object that can be used as
it's appropriate argument in \code{\link[readr:read_delim]{readr::read_delim()}} or \code{\link[readr:locale]{readr::locale()}}
}
\description{
The \code{guess} family of functions helps you figure out certain
structural characteristics of flat files without you having
to open the file and examine it manually.

What each specific function does can usually be found out by
taking a look at the corresponding \code{\link[readr:read_delim]{readr::read_delim()}} argument or
by running \code{?guess_*} and reading the "Details" section.
}
\details{
Here is a list of all currently available \code{guess} functions and their respective
descriptions:
\itemize{
\item \code{\link[=guess_col_names]{guess_col_names()}}: guesses the column names (if the file has a header)
\item \code{\link[=guess_col_types]{guess_col_types()}}: guesses the column types
\item \code{\link[=guess_decimal_mark]{guess_decimal_mark()}}: guesses the decimal mark used in the file
\item \code{\link[=guess_delim]{guess_delim()}}: guesses the file's delimiter
\item \code{\link[=guess_encoding]{guess_encoding()}}: guesses the file's encoding
\item \code{\link[=guess_grouping_mark]{guess_grouping_mark()}}: guesses the grouping/thousands mark used in the file
\item \code{\link[=guess_has_header]{guess_has_header()}}: gusses whether the file has column names
\item \code{\link[=guess_quote]{guess_quote()}}: gusses the character used to quote strings (if strings are quoted)
\item \code{\link[=guess_skip]{guess_skip()}}: guesses how many blank lines are at the beginning of the file
}
}
\examples{
\dontrun{
# Create sample file
file <- tempfile()
write.table(
  dplyr::storms, file, sep = "|", dec = ",",
  col.names = TRUE, row.names = FALSE)
  
# Run some guesses independetly
guess_delim(file)
guess_decimal_mark(file)
guess_col_names(file)
}
}
\seealso{
\code{\link[readr:read_delim]{readr::read_delim()}}, \code{\link[readr:locale]{readr::locale()}}
}
