% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_aux.R
\name{raw_to_char}
\alias{raw_to_char}
\title{Auxiliar function to convert vector of raws to chars}
\usage{
raw_to_char(raw)
}
\arguments{
\item{raw}{Vector of raw characters}
}
\description{
Auxiliar function to convert vector of raws to chars
}
