% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_accent}
\alias{rm_accent}
\title{Remove accents from words}
\usage{
rm_accent(str)
}
\arguments{
\item{str}{a character vector.}
}
\description{
Replace each character with it accentless version, including cedillas.

Current accents:

acute = áéíóúÁÉÍÓÚýÝ
grave = àèìòùÀÈÌÒÙ
circunflex = âêîôûÂÊÎÔÛ
tilde = "ãõÃÕñÑ
umlaut = äëïöüÄËÏÖÜÿ
cedil = çÇ
}
\examples{

rm_accent(c("àquela", "Ñ", "opção", "açaí"))

}
