% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_aux.R
\name{safe_read}
\alias{safe_read}
\title{Auxiliar function to safely read lines}
\usage{
safe_read(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[readr:read_lines]{readr::read_lines()}}}
}
\description{
Auxiliar function to safely read lines
}
