% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tide_names}
\alias{tide_names}
\title{Tide up names to fit as column names of a data.frame}
\usage{
tide_names(x)
}
\arguments{
\item{x}{a character vector.}
}
\description{
Sanitize a character vector to fit as column names.
}
\examples{

tide_names(names(iris))
tide_names(c("repeated_name", "Repeated___NAME", "RePeated...nAMe  "))
tide_names(c("snakeCase", "snake_case", "snake CasE", "snake Cas E"))

}
