% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frk.R
\name{frk_summarise}
\alias{frk_summarise}
\title{Guess all structural characteristics of one or more flat files}
\usage{
frk_summarise(path, pattern = NULL, recursive = FALSE, guess_max = 10,
  skip = NULL, verbose = FALSE, progress = TRUE)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the working directory, getwd(). Tilde expansion (see path.expand) is performed. Missing values will be ignored.}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be returned.}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{guess_max}{Maximum number of records to use for guess}

\item{skip}{atomic numeric specifying the number of lines to skip in each file. If NULL, it will be guessed.}

\item{verbose}{Whether to output guess as message}

\item{progress}{Show progress bar?}
}
\value{
A tibble whose lines stores the guesses for each file
}
\description{
Runs all functions from the \code{guess} family to gather all structural
characteristics of given file(s).
}
\examples{
\dontrun{
# Create sample files
file <- tempfile()
write.table(
  dplyr::storms, file, sep = "|", dec = ",",
  col.names = TRUE, row.names = FALSE)
file2 <- tempfile()
write_csv(dplyr::storms, file2)

# Run all guesses
summ <- frk_summarise(file)
str(summ)

# Run all guesses on multiple files
summ <- frk_summarise(c(file, file2))
str(summ)
}

}
\seealso{
\code{\link[=guess_delim]{guess_delim()}}, \code{\link[=guess_col_names]{guess_col_names()}}, and the whole \code{guess} family
}
