% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_aux.R
\name{read_with_guess}
\alias{read_with_guess}
\title{Auxiliar function to read file while guessing delimiter and skip}
\usage{
read_with_guess(file, guess_max)
}
\arguments{
\item{file}{Path to file}

\item{guess_max}{Maximum number of records to use for guesses}
}
\description{
Auxiliar function to read file while guessing delimiter and skip
}
